+++ Verblud_fast.h
@@ -6,12 +6,12 @@ public:
 
     Verblud_fast();
 
-    int metod_finish(const int distance) override;
+    double metod_finish(const int distance) override;
 
 protected:
     Verblud_fast(const std::string &name,const int speed,const int stamina);
 
-    int metod_relax(const int distance, const int times); //метод вычисления общего времени остановок
+    double metod_relax(const int distance, const int times); //метод вычисления общего времени остановок
 
 };
